Improve visuals in <SignUp /> & <SignIn />
