+++ frontend/src/routes/SignIn/styles.module.css
@@ -13,12 +13,3 @@
   color: black;
   margin-top: 0;
 }
-
-.status {
-  color: grey;
-  margin: 0;
-}
-
-.status.error {
-  color: red;
-}
