+++ frontend/src/routes/SignUp/index.js
@@ -6,7 +6,6 @@ import CssBaseline from '@material-ui/core/CssBaseline';
 import TextField from '@material-ui/core/TextField';
 import Link from '@material-ui/core/Link';
 import Grid from '@material-ui/core/Grid';
-import Box from '@material-ui/core/Box';
 import LockOutlinedIcon from '@material-ui/icons/LockOutlined';
 import { makeStyles } from '@material-ui/core/styles';
 import Container from '@material-ui/core/Container';
@@ -14,9 +13,10 @@ import { useDispatch, useSelector } from 'react-redux';
 import { useState } from 'react';
 
 import {signUp} from 'actions/auth';
-import { Copyright } from 'components/Copyright';
+import Copyright from 'components/Copyright';
 import { ERROR } from 'actions/types';
 import styles from './styles.module.css'
+import paths from 'routes/paths';
 
 
 
@@ -42,10 +42,7 @@ const useStyles = makeStyles((theme) => ({
     maxWidth: '350px',
     background:'white',
     padding: '0 2rem 2rem 2rem',
-    position:'absolute',
-    top: 'calc(50% - 520px/2)',
-    left: 'calc(50% - 350px/2)',
-    height: '520px'
+    position: 'absolute',
   }
 }));
 
@@ -67,98 +64,97 @@ export default function SignUp(props) {
     event.preventDefault();
     dispatch(signUp(formValues))
   };
-
   
   const status = useSelector(state => state.auth.status);
 
   return (
-    <Container className={classes.container} component="main">
-      <CssBaseline />
-      <div className={classes.paper}>
-        <Avatar className={classes.avatar}>
-          <LockOutlinedIcon />
-        </Avatar>
-        <h1 className={styles.title}>Sign Up</h1>
-        {status.detail && (
-          <p className={classNames(styles.status, status.type === ERROR && styles.error)}>
-            {status.detail}
-          </p>
-        )}
-        <form className={classes.form} noValidate>
-          <Grid container spacing={2}>
-            <Grid item xs={12}>
-              <TextField
-                autoComplete="fname"
-                name="firstName"
-                variant="outlined"
-                required
-                fullWidth
-                id="firstName"
-                label="First Name"
-                autoFocus
-                onChange={handleFormChange('firstName')}
-              />
+    <div className={styles.root}>
+      <Container className={classes.container} component="main">
+        <CssBaseline />
+        <div className={classes.paper}>
+          <Avatar className={classes.avatar}>
+            <LockOutlinedIcon />
+          </Avatar>
+          <h1 className={styles.title}>Sign Up</h1>
+          {status.detail && (
+            <p className={classNames(styles.status, status.type === ERROR && styles.error)}>
+              {status.detail}
+            </p>
+          )}
+          <form className={classes.form} noValidate>
+            <Grid container spacing={2}>
+              <Grid item xs={12}>
+                <TextField
+                  autoComplete="fname"
+                  name="firstName"
+                  variant="outlined"
+                  required
+                  fullWidth
+                  id="firstName"
+                  label="First Name"
+                  autoFocus
+                  onChange={handleFormChange('firstName')}
+                />
+              </Grid>
+              <Grid item xs={12}>
+                <TextField
+                  variant="outlined"
+                  required
+                  fullWidth
+                  id="username"
+                  label="Username"
+                  name="username"
+                  autoComplete="username"
+                  onChange={handleFormChange('username')}
+                />
+              </Grid>
+              <Grid item xs={12}>
+                <TextField
+                  variant="outlined"
+                  required
+                  fullWidth
+                  id="email"
+                  label="Email Address"
+                  name="email"
+                  autoComplete="email"
+                  onChange={handleFormChange('email')}
+                />
+              </Grid>
+              <Grid item xs={12}>
+                <TextField
+                  variant="outlined"
+                  required
+                  fullWidth
+                  name="password"
+                  label="Password"
+                  type="password"
+                  id="password"
+                  autoComplete="current-password"
+                  onChange={handleFormChange('password')}
+                />
+              </Grid>
             </Grid>
-            <Grid item xs={12}>
-              <TextField
-                variant="outlined"
-                required
-                fullWidth
-                id="username"
-                label="Username"
-                name="username"
-                autoComplete="username"
-                onChange={handleFormChange('username')}
-              />
+            <Button
+              type="submit"
+              fullWidth
+              variant="contained"
+              color="primary"
+              className={classes.submit}
+              onClick={handleSubmit}
+            >
+              Sign Up
+            </Button>
+            <Grid container justify="flex-end">
+              <Grid item>
+                <Link onClick={()=>props.history.push(paths.signIn)} variant="body2">
+                  Already have an account? Sign in
+                </Link>
+              </Grid>
             </Grid>
-            <Grid item xs={12}>
-              <TextField
-                variant="outlined"
-                required
-                fullWidth
-                id="email"
-                label="Email Address"
-                name="email"
-                autoComplete="email"
-                onChange={handleFormChange('email')}
-              />
-            </Grid>
-            <Grid item xs={12}>
-              <TextField
-                variant="outlined"
-                required
-                fullWidth
-                name="password"
-                label="Password"
-                type="password"
-                id="password"
-                autoComplete="current-password"
-                onChange={handleFormChange('password')}
-              />
-            </Grid>
-          </Grid>
-          <Button
-            type="submit"
-            fullWidth
-            variant="contained"
-            color="primary"
-            className={classes.submit}
-            onClick={handleSubmit}
-          >
-            Sign Up
-          </Button>
-          <Grid container justify="flex-end">
-            <Grid item>
-              <Link onClick={()=>props.history.push('/signin')} variant="body2">
-                Already have an account? Sign in
-              </Link>
-            </Grid>
-          </Grid>
-        </form>
-      </div>
-      <Box mt={5}>
+          </form>
+        </div>
         <Copyright />
-      </Box>
-    </Container>
+      </Container>
+    </div>
   );
 }
